Create Walmart Question
Suppose you are given a tree in which there are nodes from 1-n (rooted at 1). And there are 3 operations available which -: ID1 to color all nodes in the subtree as red excluding the parent node, ID2 to color all nodes in the subtree as black excluding the parent node and ID3 count all the red nodes in the subtree excluding the parent node.