Create shortest_k_path.cpp
This is algorithm is based on the idea of Dijkstra's shortest path algorithm. Using it first find the shortest path, then remove this path and again find the shortest path until k.
If does not pass any test case please do comment out. And more optimised solution is always welcome